+++ CSharp/Concentus/Opus/Structs/ChannelLayout.cs
@@ -47,14 +47,14 @@ namespace Concentus.Structs
         internal int nb_channels;
         internal int nb_streams;
         internal int nb_coupled_streams;
-        internal readonly byte[] mapping = new byte[256];
+        internal readonly short[] mapping = new short[256];
 
         internal void Reset()
         {
             nb_channels = 0;
             nb_streams = 0;
             nb_coupled_streams = 0;
-            Arrays.MemSet<byte>(mapping, 0);
+            Arrays.MemSet<short>(mapping, 0);
         }
     }
 }
